Add process IDs to .pidfile in start.sh and handle file existence in stop.sh

+++ Jenkins/stop.sh
@@ -1,1 +1,2 @@
-kill $(cat .pidfile)
+if [ -f .pidfile ]; then
+    kill $(cat .pidfile)
